glcli: Rename package
In ordert to stick to the naming scheme that has established around
greenlight crates which is `gl-<package>` we change the name of the
crate to be published to crates.io. When using the binary, the newly
introduced hyphen might be a bit cumbersome so we keep the name of the
binary `glcli`.
